refactor(FE): 유틸리티 함수 모듈화 및 코드 정리
- `handleEnterKeyDown` 및 `createPreviewImages` 함수를 `utils.ts`로 이동하여 코드 중복 제거 및 모듈화
- `ReviewForm.tsx`에서 해당 함수들을 `utils.ts`에서 import하여 사용하도록 수정
